+++ android/src/main/kotlin/com/istornz/live_activities/LiveActivitiesPlugin.kt
@@ -1,4 +1,4 @@
-package com.example.live_activities
+package com.istornz.live_activities
 
 import android.content.Context
 import io.flutter.embedding.engine.plugins.FlutterPlugin
@@ -38,11 +38,31 @@ class LiveActivitiesPlugin : FlutterPlugin, MethodCallHandler {
 
                 "createActivity" -> {
                     val timestamp = System.currentTimeMillis()
+                  
                     val id = args["activityId"] as String
+                    val tag = args["activityTag"] as String?
 
                     pluginScope.launch {
                         val notificationId = liveActivityManager.createActivity(
                             id,
+                            tag,
+                            timestamp,
+                            data
+                        )
+                        result.success(notificationId)
+                    }
+                }
+
+                "createOrUpdateActivity" -> {
+                    val timestamp = System.currentTimeMillis()
+
+                    val id = args["activityId"] as String
+                    val tag = args["activityTag"] as String?
+
+                    pluginScope.launch {
+                        val notificationId = liveActivityManager.createOrUpdateActivity(
+                            id,
+                            tag,
                             timestamp,
                             data
                         )
@@ -53,17 +73,19 @@ class LiveActivitiesPlugin : FlutterPlugin, MethodCallHandler {
                 "updateActivity" -> {
                     val timestamp = System.currentTimeMillis()
                     val id = args["activityId"] as String
+                    val tag = args["activityTag"] as String?
 
                     pluginScope.launch {
-                        liveActivityManager.updateActivity(id, timestamp, data)
+                        liveActivityManager.updateActivity(id, tag, timestamp, data)
                         result.success("activity/updated")
                     }
                 }
 
                 "endActivity" -> {
                     val id = args["activityId"] as String
+                    val tag = args["activityTag"] as String?
 
-                    liveActivityManager.endActivity(id, data)
+                    liveActivityManager.endActivity(id, tag, data)
                     result.success("activity/ended")
                 }