+++ test/live_activities_test.dart
@@ -19,15 +19,17 @@ class MockLiveActivitiesPlatform
   @override
   Future<String?> createActivity(
     String activityId,
+    String? activityTag,
     Map<String, dynamic> data, {
     bool removeWhenAppIsKilled = false,
+    bool iOSEnableRemoteUpdates = true,
     Duration? staleIn,
   }) {
     return Future.value('ACTIVITY_ID');
   }
 
   @override
-  Future endActivity(String activityId) {
+  Future endActivity(String activityId, String? activityTag) {
     return Future.value();
   }
 
@@ -53,15 +55,17 @@ class MockLiveActivitiesPlatform
 
   @override
   Stream<UrlSchemeData> urlSchemeStream() {
-    return Stream.value(UrlSchemeData(
-      url: 'URL',
-      scheme: 'SCHEME',
-      host: 'HOST',
-      path: 'PATH',
-      queryParameters: [
-        {'name': 'NAME', 'value': 'VALUE'},
-      ],
-    ));
+    return Stream.value(
+      UrlSchemeData(
+        url: 'URL',
+        scheme: 'SCHEME',
+        host: 'HOST',
+        path: 'PATH',
+        queryParameters: [
+          {'name': 'NAME', 'value': 'VALUE'},
+        ],
+      ),
+    );
   }
 
   @override
@@ -85,8 +89,12 @@ class MockLiveActivitiesPlatform
   }
 
   @override
-  Future updateActivity(String activityId, Map<String, dynamic> data,
-      [AlertConfig? alertConfig]) {
+  Future updateActivity(
+    String activityId,
+    String? activityTag,
+    Map<String, dynamic> data, [
+    AlertConfig? alertConfig,
+  ]) {
     return Future.value();
   }
 
@@ -98,8 +106,10 @@ class MockLiveActivitiesPlatform
   @override
   Future createOrUpdateActivity(
     String customId,
+    String? activityTag,
     Map<String, dynamic> data, {
     bool removeWhenAppIsKilled = false,
+    bool iOSEnableRemoteUpdates = true,
     Duration? staleIn,
   }) {
     return Future.value();
@@ -179,10 +189,7 @@ void main() {
   });
 
   test('getPushToken', () async {
-    expect(
-      await liveActivitiesPlugin.getPushToken('PUSH_TOKEN'),
-      'PUSH_TOKEN',
-    );
+    expect(await liveActivitiesPlugin.getPushToken('PUSH_TOKEN'), 'PUSH_TOKEN');
   });
 
   test('activityUpdateStream', () async {
@@ -205,8 +212,9 @@ void main() {
 
     expect(wrongMappingIsNull, null);
 
-    final correctMappingNotNull =
-        result.mapOrNull(active: (state) => state.activityToken);
+    final correctMappingNotNull = result.mapOrNull(
+      active: (state) => state.activityToken,
+    );
 
     expect(correctMappingNotNull, 'ACTIVITY_TOKEN');
   });